+++ packages/isaacscript-common/src/types/AnyGridEntity.ts
@@ -1,0 +1,9 @@
+export type AnyGridEntity =
+  | GridEntity
+  | GridEntityDoor
+  | GridEntityPit
+  | GridEntityPoop
+  | GridEntityPressurePlate
+  | GridEntityRock
+  | GridEntitySpikes
+  | GridEntityTNT;